requirements.txt
pytest==8.4.1
pydantic==2.11.7
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlalchemy==2.0.41
sqlmodel==0.0.24
openpyxl==3.1.5
pandas==2.3.0
python-dateutil==2.9.0
